correction on precision parameter/data typ

+++ python/remote_client_qt.py
@@ -39,7 +39,7 @@ class remote_client_qt(plotter_base):
         if precision:
             self.data_type = np.float16
         else:
-            self.data_type = np.float8
+            self.data_type = np.int8
 
         if show_axes:
             self.toggle_axes()
@@ -87,10 +87,10 @@ class remote_client_qt(plotter_base):
 
     def set_precision(self, precision):
         if precision:
-            self.data_type = np.float32
+            self.data_type = np.float16
             print '-->Remote: 16bit FFT in use (more bandwidth and precision)'
         else:
-            self.data_type = np.float16
+            self.data_type = np.int8
             print '-->Remote: 8bit FFT in use (less bandwidth and precision)'
 
     def set_hold_max(self, hold_max):
